Move INodeVersionInfoProvider to core so it could be used in telemetry
